Implement async function to fetch all posts
The getAllPosts function sorts the posts by date after having retrieved all the posts. If there are a large number of posts, this could be costly in terms of performance. A more efficient approach would be to read and sort the files one by one while writing the results to an array. This way we only need to keep one post in memory at a time, which significantly reduces memory usage.